# Conflicts:
#	instrumentation/aws-sdk/aws-sdk-2.2/testing/src/main/java/io/opentelemetry/instrumentation/awssdk/v2_2/AbstractAws2SqsTracingTest.java

+++ instrumentation/rxjava/rxjava-common-3.0/testing/src/main/java/io/opentelemetry/instrumentation/rxjava/common/v3_0/AbstractRxJava3Test.java
@@ -873,18 +873,16 @@ public abstract class AbstractRxJava3Test {
                   span ->
                       span.hasName("outer")
                           .hasNoParent()
-                          .hasAttributesSatisfyingExactly(
-                              equalTo(longKey("iteration"), (long) iteration)),
+                          .hasAttributesSatisfyingExactly(equalTo(longKey("iteration"), iteration)),
                   span ->
                       span.hasName("middle")
                           .hasParent(trace.getSpan(0))
-                          .hasAttributesSatisfyingExactly(
-                              equalTo(longKey("iteration"), (long) iteration)),
+                          .hasAttributesSatisfyingExactly(equalTo(longKey("iteration"), iteration)),
                   span ->
                       span.hasName("inner")
                           .hasParent(trace.getSpan(1))
                           .hasAttributesSatisfyingExactly(
-                              equalTo(longKey("iteration"), (long) iteration)));
+                              equalTo(longKey("iteration"), iteration)));
     }
     testing()
         .waitAndAssertSortedTraces(
